refactor(workbench): W5 定案 npm workspaces + W6 内置台真抽包
W5：暂不切 pnpm，官方路径保持 npm workspaces。
W6：local/ssh/companion descriptor/prompt（及 CompanionWorkbench）
迁入 packages；registry 全改包名；src 薄 re-export。SDK 仍 re-export。
