+++ packages/README.md
@@ -1,30 +1,25 @@
 # Packages（OEM / Workbench Monorepo）
 
-> **状态（2026-07-13）**：npm workspaces ✅；`shared-types` / `workbench-assistant` / `workbench-sample` 已落地。  
+> **状态（2026-07-13）**：npm workspaces ✅（W5：暂不切 pnpm）；内置工作台包已真抽。  
 > 完整 TODO：[`docs/workbench-monorepo-design.md` §6.0](../docs/workbench-monorepo-design.md)。
 
 ## 当前事实
 
 | 包名 | 状态 |
 |---|---|
-| `@sailfish/shared-types` | ✅ 真相源在 `packages/shared-types/src/`；`shared/types/*` 兼容 re-export |
-| `@sailfish/workbench-assistant` | ✅ descriptor / prompt / agent-tools / `AssistantWorkbench.vue`；AiPanel + artifact 仍 `@/` 引用 desktop（P2 前） |
-| `@sailfish/workbench-sample` | ✅ 业务台样例：无对话 UI + skills + 假 MCP（默认 disabled）+ bootstrap 单测 |
-| `@sailfish/workbench-sdk` 等 | ⚠️ 仍 re-export → `src/workbench`（W6） |
+| `@sailfish/shared-types` | ✅ 真相源在 `packages/shared-types/src/` |
+| `@sailfish/workbench-assistant` | ✅ descriptor / prompt / agent-tools / Vue；AiPanel+artifact 仍 `@/` |
+| `@sailfish/workbench-local` / `ssh` | ✅ descriptor / prompt；`TerminalTabView` 仍 desktop（Teleport） |
+| `@sailfish/workbench-companion` | ✅ descriptor / prompt / CompanionWorkbench；AiPanel 仍 `@/` |
+| `@sailfish/workbench-sample` | ✅ 业务台样例 + bootstrap 单测 |
+| `@sailfish/workbench-sdk` | ⚠️ 仍 re-export → `src/workbench`（types/registry 未迁） |
 
 ## 导入约定
 
-优先：
-
 ```ts
-import type { McpServerConfig, TerminalType } from '@sailfish/shared-types'
-import { descriptor } from '@sailfish/workbench-assistant'
-import { ASSISTANT_WORKBENCH_AGENT_TOOLS } from '@sailfish/workbench-assistant/agent-tools'
+import type { TerminalType } from '@sailfish/shared-types'
+import { descriptor } from '@sailfish/workbench-local/descriptor'
+import { LOCAL_WORKBENCH_AGENT_PROMPT } from '@sailfish/workbench-local/prompt'
 ```
 
-兼容（仍可用）：
-
-```ts
-import type { TerminalType } from '@shared/types'
-import { AGENT_PROMPT } from '@/workbench/assistant/prompt'
-```
+兼容：`@shared/types`、`@/workbench/<kind>/prompt` 薄 re-export 仍可用。
