+++ packages/workbench-companion/src/index.ts
@@ -1,2 +1,7 @@
-export { descriptor } from '../../../src/workbench/companion/descriptor'
-export { COMPANION_WORKBENCH_AGENT_PROMPT } from '../../../src/workbench/companion/prompt'
+/**
+ * @sailfish/workbench-companion
+ * CompanionWorkbench 在本包；AiPanel 经 `@/` 引用 desktop（P2 前）。
+ */
+export { descriptor } from './descriptor'
+export { COMPANION_WORKBENCH_AGENT_PROMPT } from './prompt'
+export { default as CompanionWorkbench } from './CompanionWorkbench.vue'
