+++ packages/workbench-companion/src/prompt.ts
@@ -1,0 +1,8 @@
+/**
+ * 联络（companion）工作台 → Agent system prompt 片段
+ *
+ * 联络只含聊天区，没有产出物面板等界面能力，因此**当前不注入任何工作台 UI prompt**
+ * （descriptor.agentPrompt 为 undefined）。
+ */
+
+export const COMPANION_WORKBENCH_AGENT_PROMPT: string | undefined = undefined
