+++ packages/workbench-local/src/index.ts
@@ -1,2 +1,6 @@
-export { descriptor } from '../../../src/workbench/local/descriptor'
-export { LOCAL_WORKBENCH_AGENT_PROMPT } from '../../../src/workbench/local/prompt'
+/**
+ * @sailfish/workbench-local
+ * TerminalTabView 仍在 desktop（Teleport）；descriptor 经 `@/` 引用。
+ */
+export { descriptor } from './descriptor'
+export { LOCAL_WORKBENCH_AGENT_PROMPT } from './prompt'
