+++ packages/workbench-ssh/package.json
@@ -2,8 +2,16 @@
   "name": "@sailfish/workbench-ssh",
   "version": "0.1.0",
   "private": true,
+  "description": "SSH 终端工作台（descriptor / prompt；TerminalTabView 在 desktop）",
   "type": "module",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
-  "exports": { ".": "./src/index.ts" }
+  "exports": {
+    ".": "./src/index.ts",
+    "./descriptor": "./src/descriptor.ts",
+    "./prompt": "./src/prompt.ts"
+  },
+  "peerDependencies": {
+    "vue": "^3.4.0"
+  }
 }
