+++ packages/workbench-ssh/src/index.ts
@@ -1,2 +1,6 @@
-export { descriptor } from '../../../src/workbench/ssh/descriptor'
-export { SSH_WORKBENCH_AGENT_PROMPT } from '../../../src/workbench/ssh/prompt'
+/**
+ * @sailfish/workbench-ssh
+ * TerminalTabView 仍在 desktop（Teleport）；descriptor 经 `@/` 引用。
+ */
+export { descriptor } from './descriptor'
+export { SSH_WORKBENCH_AGENT_PROMPT } from './prompt'
