+++ src/workbench/companion/descriptor.ts
@@ -1,11 +1,2 @@
-import CompanionWorkbench from '../../components/workbench/CompanionWorkbench.vue'
-import type { WorkbenchDescriptor } from '../types'
-import { COMPANION_WORKBENCH_AGENT_PROMPT } from './prompt'
-
-export const descriptor: WorkbenchDescriptor = {
-  kind: 'companion',
-  renderer: CompanionWorkbench,
-  // 与 assistant 一致：助手类工作台在 Steam 版不提供，回退到终端渲染器
-  availableInSteam: false,
-  agentPrompt: COMPANION_WORKBENCH_AGENT_PROMPT,
-}
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-companion/descriptor` */
+export { descriptor } from '@sailfish/workbench-companion/descriptor'
