+++ src/workbench/local/descriptor.ts
@@ -1,10 +1,2 @@
-import TerminalTabView from '../../components/TerminalTabView.vue'
-import type { WorkbenchDescriptor } from '../types'
-import { LOCAL_WORKBENCH_AGENT_PROMPT } from './prompt'
-
-export const descriptor: WorkbenchDescriptor = {
-  kind: 'local',
-  renderer: TerminalTabView,
-  availableInSteam: true,
-  agentPrompt: LOCAL_WORKBENCH_AGENT_PROMPT,
-}
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-local/descriptor` */
+export { descriptor } from '@sailfish/workbench-local/descriptor'
