+++ src/workbench/local/prompt.ts
@@ -1,16 +1,2 @@
-/**
- * local 终端工作台 → Agent system prompt 片段
- *
- * 包含本地 PTY 终端操作规范，注入到 AgentContext.workbenchPrompt。
- */
-
-export const LOCAL_WORKBENCH_AGENT_PROMPT = [
-  '# 本地终端工作台',
-  '当前对话在**本地终端工作台**中进行。左侧是 PTY 终端，Agent 可直接执行本地命令、读写本地文件。',
-  '**禁止的命令**：vim/vi/nano/emacs（用 `write_text_file`）、tmux/screen、mc/ranger',
-  '**长内容处理**：',
-  '- 超过 200 字符禁止用 echo/printf，用 `write_text_file` 写入 /tmp 再执行',
-  '- 长文本分析结果直接在对话中回复，不要发送到终端',
-  '**长耗时命令**：执行 → `wait` 等待 → `check_terminal_status` 确认，超时不代表失败',
-  '- 等待时可以说点有趣的话，比如："去喝杯咖啡☕马上回来"、"编译中，先摸会儿鱼🐟"、"让子弹飞一会儿🎬"',
-].join('\n')
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-local/prompt` */
+export { LOCAL_WORKBENCH_AGENT_PROMPT } from '@sailfish/workbench-local/prompt'
