+++ src/workbench/registry.ts
@@ -11,10 +11,10 @@ import { isOemFeatureEnabled, type OemFeatureKey } from '@shared/oem-features'
 import TerminalTabView from '../components/TerminalTabView.vue'
 import type { WorkbenchKind } from './types'
 import { getWorkbenchDescriptor, registerWorkbench } from './registry-store'
-import { descriptor as localDescriptor } from './local/descriptor'
-import { descriptor as sshDescriptor } from './ssh/descriptor'
+import { descriptor as localDescriptor } from '@sailfish/workbench-local/descriptor'
+import { descriptor as sshDescriptor } from '@sailfish/workbench-ssh/descriptor'
 import { descriptor as assistantDescriptor } from '@sailfish/workbench-assistant/descriptor'
-import { descriptor as companionDescriptor } from './companion/descriptor'
+import { descriptor as companionDescriptor } from '@sailfish/workbench-companion/descriptor'
 import { descriptor as sampleDescriptor } from '@sailfish/workbench-sample/descriptor'
 
 export {
