+++ src/workbench/ssh/descriptor.ts
@@ -1,10 +1,2 @@
-import TerminalTabView from '../../components/TerminalTabView.vue'
-import type { WorkbenchDescriptor } from '../types'
-import { SSH_WORKBENCH_AGENT_PROMPT } from './prompt'
-
-export const descriptor: WorkbenchDescriptor = {
-  kind: 'ssh',
-  renderer: TerminalTabView,
-  availableInSteam: true,
-  agentPrompt: SSH_WORKBENCH_AGENT_PROMPT,
-}
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-ssh/descriptor` */
+export { descriptor } from '@sailfish/workbench-ssh/descriptor'
