+++ src/workbench/ssh/prompt.ts
@@ -1,29 +1,2 @@
-/**
- * ssh 终端工作台 → Agent system prompt 片段
- *
- * 包含 SSH 远程终端操作规范，注入到 AgentContext.workbenchPrompt。
- */
-
-export const SSH_WORKBENCH_AGENT_PROMPT = [
-  '# SSH 远程终端工作台',
-  '',
-  '当前对话在 **SSH 远程终端工作台**中进行，Agent 操作的是**远程服务器**，不是本地机器。',
-  '',
-  '**关键限制**：',
-  '- `read_file`、`edit_file`、`write_text_file` **不可用**（只能操作本地文件）',
-  '- 读取远程文件用 `cat`/`head`/`tail`，写入用 `write_remote_text_file` 或 `echo`/`cat <<EOF`',
-  '- 终端状态需根据屏幕内容自行判断（看提示符、Password:、进度等）',
-  '',
-  '**禁止的命令**：vim/vi/nano/emacs（用 `write_remote_text_file`）、tmux/screen、mc/ranger',
-  '',
-  '**长内容处理**：超过 200 字符用 `write_remote_text_file` 写入 /tmp 再执行，禁止 echo/printf',
-  '',
-  '**长耗时命令**：执行 → `wait` 等待 → `check_terminal_status` 确认，超时不代表失败',
-  '- 等待时可以说点有趣的话，比如："去喝杯咖啡☕马上回来"、"编译中，先摸会儿鱼🐟"、"让子弹飞一会儿🎬"',
-  '',
-  '**SSH 终端状态判断**（根据屏幕内容）：',
-  '- 看到 `$` 或 `#` 提示符 → 可执行新命令',
-  '- 看到 `Password:` → 暂停，让用户输入',
-  '- 看到 `(y/n)` → 根据情况回复或询问用户',
-  '- 看到 `--More--` 或 `(END)` → 发送 `q` 退出',
-].join('\n')
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-ssh/prompt` */
+export { SSH_WORKBENCH_AGENT_PROMPT } from '@sailfish/workbench-ssh/prompt'
