+++ tsconfig.json
@@ -28,8 +28,11 @@
       "@sailfish/workbench-sample": ["./packages/workbench-sample/src/index.ts"],
       "@sailfish/workbench-sample/*": ["./packages/workbench-sample/src/*"],
       "@sailfish/workbench-local": ["./packages/workbench-local/src/index.ts"],
+      "@sailfish/workbench-local/*": ["./packages/workbench-local/src/*"],
       "@sailfish/workbench-ssh": ["./packages/workbench-ssh/src/index.ts"],
-      "@sailfish/workbench-companion": ["./packages/workbench-companion/src/index.ts"]
+      "@sailfish/workbench-ssh/*": ["./packages/workbench-ssh/src/*"],
+      "@sailfish/workbench-companion": ["./packages/workbench-companion/src/index.ts"],
+      "@sailfish/workbench-companion/*": ["./packages/workbench-companion/src/*"]
     }
   },
   "include": [
@@ -41,7 +44,11 @@
     "packages/workbench-assistant/src/**/*.ts",
     "packages/workbench-assistant/src/**/*.vue",
     "packages/workbench-sample/src/**/*.ts",
-    "packages/workbench-sample/src/**/*.vue"
+    "packages/workbench-sample/src/**/*.vue",
+    "packages/workbench-local/src/**/*.ts",
+    "packages/workbench-ssh/src/**/*.ts",
+    "packages/workbench-companion/src/**/*.ts",
+    "packages/workbench-companion/src/**/*.vue"
   ],
   "exclude": ["src/**/__tests__/**", "src/**/*.test.ts"]
 }
