+++ vite.config.ts
@@ -314,6 +314,9 @@ export default defineConfig({
       '@sailfish/workbench-sdk': resolve(__dirname, 'packages/workbench-sdk/src/index.ts'),
       '@sailfish/workbench-assistant': resolve(__dirname, 'packages/workbench-assistant/src'),
       '@sailfish/workbench-sample': resolve(__dirname, 'packages/workbench-sample/src'),
+      '@sailfish/workbench-local': resolve(__dirname, 'packages/workbench-local/src'),
+      '@sailfish/workbench-ssh': resolve(__dirname, 'packages/workbench-ssh/src'),
+      '@sailfish/workbench-companion': resolve(__dirname, 'packages/workbench-companion/src'),
     }
   },
   build: {
