+++ vitest.config.ts
@@ -22,6 +22,9 @@ export default defineConfig({
       '@sailfish/shared-types': resolve(__dirname, 'packages/shared-types/src/index.ts'),
       '@sailfish/workbench-assistant': resolve(__dirname, 'packages/workbench-assistant/src'),
       '@sailfish/workbench-sample': resolve(__dirname, 'packages/workbench-sample/src'),
+      '@sailfish/workbench-local': resolve(__dirname, 'packages/workbench-local/src'),
+      '@sailfish/workbench-ssh': resolve(__dirname, 'packages/workbench-ssh/src'),
+      '@sailfish/workbench-companion': resolve(__dirname, 'packages/workbench-companion/src'),
     }
   }
 })
